feat(#2): refatorações para componentizar o app e separar melhor cada coisa
