+++ src/mocks/cesta.js
@@ -1,0 +1,17 @@
+import topo from '../../assets/topo.png'
+import logo from '../../assets/logo.png'
+const cesta = {
+    topo: {
+        titulo: "Detalhe da cesta",
+        imagem:topo,
+    },
+    detalhes: {
+        nome: "Cesta de verduras",
+        imagem:logo,
+        nomeFazenda: "Jenny Jack Farm",
+        descricao:"Uma cesta com produtos selecionados cuidadosamente  da fazenda direto para a sua cozinha",
+        preco:"R$ 40,00",
+    }  
+}
+
+export default cesta;
